// src/interfaces/IApgRpnFragmentifyResult.ts
/** -----------------------------------------------------------------------
 * @module [RPN]
 * @author [APG] ANGELI Paolo Giusto
 * @version 0.8.0 [APG 2022/05/24]
 * -----------------------------------------------------------------------
 */

import { IApgRpnFragment } from "./IApgRpnFragment.ts";


export interface IApgRpnFragmentifyResult {

    fragments: IApgRpnFragment[];
    log: string[];

}